Move helper function to non-temporary file
